+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "geopolars-arrow"
+name = "geoarrow"
 version = "0.0.1"
 authors = ["Kyle Barron <kylebarron2@gmail.com>"]
 edition = "2021"
