+++ README.md
@@ -1,0 +1,3 @@
+# `geoarrow`
+
+A Rust implementation of the [GeoArrow](https://github.com/geoarrow/geoarrow) specification.
