+++ src/trait_.rs
@@ -7,8 +7,10 @@ pub trait GeometryArrayTrait<'a> {
     type ScalarGeo: From<Self::Scalar>;
     type ArrowArray;
 
+    /// Access the value at slot `i` as an Arrow scalar, not considering validity.
     fn value(&'a self, i: usize) -> Self::Scalar;
 
+    /// Access the value at slot `i` as an Arrow scalar, considering validity.
     fn get(&'a self, i: usize) -> Option<Self::Scalar> {
         if self.is_null(i) {
             return None;
@@ -17,11 +19,12 @@ pub trait GeometryArrayTrait<'a> {
         Some(self.value(i))
     }
 
+    /// Access the value at slot `i` as a [`geo`] scalar, not considering validity.
     fn value_as_geo(&'a self, i: usize) -> Self::ScalarGeo {
         self.value(i).into()
     }
 
-    /// Gets the value at slot `i` as a geo object, additionally checking the validity bitmap
+    /// Access the value at slot `i` as a [`geo`] scalar, considering validity.
     fn get_as_geo(&'a self, i: usize) -> Option<Self::ScalarGeo> {
         if self.is_null(i) {
             return None;
@@ -30,26 +33,28 @@ pub trait GeometryArrayTrait<'a> {
         Some(self.value_as_geo(i))
     }
 
+    /// Convert this array into an [`arrow2`] array.
+    /// # Implementation
+    /// This is `O(1)`.
     fn into_arrow(self) -> Self::ArrowArray;
 
-    /// Build a spatial index containing this array's geometries
+    /// Build an [`RTree`] spatial index containing this array's geometries.
     fn rstar_tree(&'a self) -> RTree<Self::Scalar>;
 
-    /// The length of the [`GeometryArray`]. Every array has a length corresponding to the number of
-    /// elements (slots).
+    /// The number of geometries contained in this array.
     fn len(&self) -> usize;
 
-    /// whether the array is empty
+    /// Returns `true` if the array is empty.
     fn is_empty(&self) -> bool {
         self.len() == 0
     }
 
-    /// The validity of the [`GeometryArray`]: every array has an optional [`Bitmap`] that, when available
-    /// specifies whether the array slot is valid or not (null).
-    /// When the validity is [`None`], all slots are valid.
+    /// Access the array's validity. Every array has an optional [`Bitmap`] that, when available
+    /// specifies whether the array slot is valid or not (null). When the validity is [`None`], all
+    /// slots are valid.
     fn validity(&self) -> Option<&Bitmap>;
 
-    /// The number of null slots on this [`GeometryArray`].
+    /// The number of null slots in this array.
     /// # Implementation
     /// This is `O(1)` since the number of null elements is pre-computed.
     #[inline]
@@ -79,7 +84,7 @@ pub trait GeometryArrayTrait<'a> {
         !self.is_null(i)
     }
 
-    /// Slices the [`GeometryArray`], returning a new `Box<dyn GeometryArray>`.
+    /// Slices the array, returning a new geometry array of the same type.
     /// # Implementation
     /// This operation is `O(1)` over `len`, as it amounts to increase two ref counts
     /// and moving the struct to the heap.
@@ -87,7 +92,7 @@ pub trait GeometryArrayTrait<'a> {
     /// This function panics iff `offset + length > self.len()`.
     fn slice(&self, offset: usize, length: usize) -> Self;
 
-    /// Slices the [`GeometryArray`], returning a new `Box<dyn GeometryArray>`.
+    /// Slices the array, returning a new geometry array of the same type.
     /// # Implementation
     /// This operation is `O(1)` over `len`, as it amounts to increase two ref counts
     /// and moving the struct to the heap.
@@ -100,7 +105,7 @@ pub trait GeometryArrayTrait<'a> {
     // /// This function panics iff `validity.len() != self.len()`.
     // fn with_validity(&self, validity: Option<Bitmap>) -> Box<dyn GeometryArray>;
 
-    /// Clone a `&dyn GeometryArray` to an owned `Box<dyn GeometryArray>`.
+    /// Clones this array to an owned, boxed geometry array.
     fn to_boxed(&self) -> Box<Self>;
 }
